DtbLoader: refactor out helper to try loading dtb
If we are going to pick a dtb based on SMBIOS info, we'll want to try a
few possibilities, so refactor out a helper to try to load a dtb from
a path constructed "printf-style" to reduce the boilerplate path
construction logic.

This drops using relative path of DtbLoader.efi in constructing the
path, in favor of finding them relative to the root.  I'm not there is a
good argument to give distros the possibility to put things in different
locations.

It also moves the CRC calculation to RegisterDtBlob(), but this will
come in handy when we add fdt overlay/patching support.

Signed-off-by: Rob Clark <robdclark@gmail.com>