DtbLoader: add panel-id detection/fixup for qcom devices
Put this in it's own file.  Maybe at some point there will be other
aarch64 laptops from other SoC vendors which need their own specific
dt patching.

This looks for UEFIDisplayInfo EFI variable, and if found extracts the
panel-id, and attempts to patch the dtb by loading a panel specific dtb
overlay.

In theory, the panel-id appears to be a flat namespace (ie. if some
panel *where* used on multiple different laptops, it would have the same
panel-id), but we first check model specific paths to find the overlay
in case this proves not to be true some day.  The paths we look for a
panel over lay are, from highest priorty to lowest:

1) \dtb\$SysVendor\$ProductName-$BoardName-panel-$PanelId.dtb
2) \dtb\$SysVendor\$ProductName-panel-$PanelId.dtb
3) \dtb\panels\panel-$PanelId.dtb

I expect the last path to be typically where the panel overlay is loaded
from.

Signed-off-by: Rob Clark <robdclark@gmail.com>