WIP: DtbLoader: register ESRT table entry
The rough idea is to provide an ESRT table entry that a (not yet
completed) fwupd plugin can use to know that the system is using
dtbloader and is interesting in installing updated dtb tables.

NOTE: this might not be the final solution.. the capsule-update
mechanism doesn't really work for this usecase, where the fwupd
plugin really just needs to drop new dtb files into the ESP
partition.