DtbLoader: Try to pick dtb based on SMBIOS tables
To better support multiple different aarch64 laptops with a single
image, add logic to pick the dtb to load based on SMBIOS tables.

This involves jumping thru some hoops, as the SMBIOS tables may not
yet be installed when DtbLoader is installed, so we need to hook
notifiers to defer SMBIOS parsing and dtb loading.

The paths that we attempt to load dtb are, in order of priority from
highest to lowest:

1) \dtb\$SysVendor\$ProductName-$BoardName.dtb
2) \dtb\$SysVendor\$ProductName.dtb
3) \my.dtb

Signed-off-by: Rob Clark <robdclark@gmail.com>