remove original readme
Although the DtbLoader tree is based on edk2, that isn't really it's
purpose.  It is done this way since we don't have pkgconfig and shared
libs and that sort of thing.  But everything is static linked and
interfaces to UEFI are based on standardized protocol interfaces, so
there is nothing that really needs to be rebased on upstream edk2.
The only reason to rebase or merge in a newer edk2 is if DtbLoader
gained a dependency on a new UEFI protocol or edk2 util module.

And in particular we want to make room for an DtbLoader specific
readme.  So remove the edk2 one which isn't relevant to make room.

Signed-off-by: Rob Clark <robdclark@gmail.com>